Move layout and dark mode to app/global level - move last updated fetch to static props
