+++ components/layout.js
@@ -9,19 +9,8 @@ import { MdEmail } from 'react-icons/md';
 const name = 'Alec Chen';
 export const siteTitle = 'Alec Chen';
 
-export default function Layout({ children, home, skinny, selected }) {
+export default function Layout({ children, home, skinny, selected, toggleDarkMode, colorMode }) {
 
-  const [ colorMode, setColorMode ] = useState('Dark Mode');
-
-  const toggleDarkMode = () => {
-    if (document.body.classList.contains("dark-mode")) {
-      document.body.classList.remove("dark-mode");
-      setColorMode('Dark Mode');
-    } else {
-      document.body.classList.add("dark-mode");
-      setColorMode('Light Mode');
-    }
-  }
 
   return (
     <div className={styles.container}>
