+++ pages/[id].js
@@ -6,14 +6,14 @@ const postsDirectory = process.cwd() + '/main'
 
 export default function Post({ postData }) {
   return (
-  <Layout skinny>
+  <>
     <Head>
       <title>{postData.title}</title>
     </Head>
     <article>
       <div dangerouslySetInnerHTML={{ __html: postData.content }} />
     </article> 
-  </Layout>
+  </>
   )
 }
