+++ pages/index.js
@@ -9,26 +9,10 @@ import { getSortedPostsData, getPostData } from '../lib/posts';
 
 const postsDirectory = process.cwd() + '/main'
 
-export default function Home({ postData }) {
-
-  const [lastUpdated, setLastUpdated] = useState('Loading...');
-
-  useEffect(() => {
-    const getRepoData = async () => {
-      const repoData = await axios({
-        method: 'get',
-        url: 'https://api.github.com/repos/alecchendev/alecchendev.github.io',
-        headers: {'Accept': 'application/vnd.github.v3+json'}
-      });
-      const dateString = repoData.data.updated_at;
-      const formatted = format(parseISO(dateString), 'LLLL d, yyyy')
-      setLastUpdated(formatted);
-    }
-    getRepoData();
-  }, [])
+export default function Home({ postData, lastUpdated }) {
 
   return (
-    <Layout home selected={''}>
+    <>
       <Head>
         <title>{siteTitle}</title>
       </Head>
@@ -37,15 +21,25 @@ export default function Home({ postData }) {
 
       <p className={"date"}>Last updated: {lastUpdated}</p>
 
-    </Layout>
+    </>
   )
 }
 
 export async function getStaticProps() {
   const postData = await getPostData(postsDirectory, 'index');
+
+  const repoData = await axios({
+    method: 'get',
+    url: 'https://api.github.com/repos/alecchendev/alecchendev.github.io',
+    headers: {'Accept': 'application/vnd.github.v3+json'}
+  });
+  const dateString = repoData.data.updated_at;
+  const formatted = format(parseISO(dateString), 'LLLL d, yyyy')
+
   return {
     props: {
-      postData
+      postData,
+      lastUpdated: formatted
     }
   }
